move to opencensus.common namespace
