+++ contrib/opencensus-common-correlationcontext/opencensus/common/__init__.py
@@ -1,0 +1,1 @@
+__path__ = __import__('pkgutil').extend_path(__path__, __name__)
